# Optimize character conversion