--- a/.rustfmt.toml
+++ b/.rustfmt.toml
@@ -1,5 +1,4 @@
-unstable_features = true
-
 use_small_heuristics = "max"
 imports_granularity = "Module"
 group_imports = "StdExternalCrate"
+hard_tabs = true