--- a/src/lib.rs
+++ b/src/lib.rs
@@ -29,6 +29,7 @@
 //!
 //! let result_four: String = converter.simplified_to_traditional(simplified_text);
 //! println!("{}", result_four == traditional_text); // --> true
+#![cfg_attr(feature = "bench", feature(test))]
 
 extern crate bincode;
 
@@ -37,39 +38,62 @@ pub use self::character_converter::Converter as CharacterConverter;
 
 #[cfg(test)]
 mod tests {
-	use super::*;
+    use super::*;
 
-	#[test]
-	fn is_traditional() {
-		let converter: CharacterConverter = CharacterConverter::new();
-		let simplified = "欧洲";
-		let traditional = "歐洲";
-		assert_eq!(true, converter.is_traditional(traditional));
-		assert_eq!(false, converter.is_traditional(simplified));
-	}
+    #[test]
+    fn is_traditional() {
+        let converter: CharacterConverter = CharacterConverter::new();
+        let simplified = "欧洲";
+        let traditional = "歐洲";
+        assert_eq!(true, converter.is_traditional(traditional));
+        assert_eq!(false, converter.is_traditional(simplified));
+    }
 
-	#[test]
-	fn is_simplified() {
-		let converter: CharacterConverter = CharacterConverter::new();
-		let simplified = "欧洲";
-		let traditional = "歐洲";
-		assert_eq!(true, converter.is_simplified(simplified));
-		assert_eq!(false, converter.is_simplified(traditional));
-	}
+    #[test]
+    fn is_simplified() {
+        let converter: CharacterConverter = CharacterConverter::new();
+        let simplified = "欧洲";
+        let traditional = "歐洲";
+        assert_eq!(true, converter.is_simplified(simplified));
+        assert_eq!(false, converter.is_simplified(traditional));
+    }
 
-	#[test]
-	fn traditional_to_simplified() {
-		let converter: CharacterConverter = CharacterConverter::new();
-		let simplified = "欧洲";
-		let traditional = "歐洲";
-		assert_eq!(simplified, converter.traditional_to_simplified(traditional));
-	}
+    #[test]
+    fn traditional_to_simplified() {
+        let converter: CharacterConverter = CharacterConverter::new();
+        let simplified = "欧洲";
+        let traditional = "歐洲";
+        assert_eq!(simplified, converter.traditional_to_simplified(traditional));
+    }
 
-	#[test]
-	fn simplified_to_traditional() {
-		let converter: CharacterConverter = CharacterConverter::new();
-		let simplified = "欧洲";
-		let traditional = "歐洲";
-		assert_eq!(traditional, converter.simplified_to_traditional(simplified));
-	}
+    #[test]
+    fn simplified_to_traditional() {
+        let converter: CharacterConverter = CharacterConverter::new();
+        let simplified = "欧洲";
+        let traditional = "歐洲";
+        assert_eq!(traditional, converter.simplified_to_traditional(simplified));
+    }
+}
+
+#[cfg(all(feature = "bench", test))]
+mod benches {
+    extern crate test;
+    use super::*;
+    use test::Bencher;
+
+    #[bench]
+    #[cfg(feature = "bench")]
+    fn traditional_to_simplified(b: &mut Bencher) {
+        let converter: CharacterConverter = CharacterConverter::new();
+        let traditional = "人人生而自由﹐在尊嚴和權利上一律平等。他們賦有理性和良心﹐並應以兄弟關係的精神互相對待。";
+        b.iter(|| converter.traditional_to_simplified(traditional));
+    }
+
+    #[bench]
+    #[cfg(feature = "bench")]
+    fn simplified_to_traditional(b: &mut Bencher) {
+        let converter: CharacterConverter = CharacterConverter::new();
+        let simplified = "人人生而自由﹐在尊严和权利上一律平等。他们赋有理性和良心﹐并应以兄弟关系的精神互相对待。";
+        b.iter(|| converter.simplified_to_traditional(simplified));
+    }
 }