--- a/src/lib.rs
+++ b/src/lib.rs
@@ -34,12 +34,25 @@ extern crate bincode;
 use std::collections::HashMap;
 
 use bincode::deserialize_from;
+use fst::raw::{Fst, Output};
 use once_cell::sync::Lazy;
 
 static T2S: Lazy<HashMap<String, String>> =
     Lazy::new(|| deserialize_from(&include_bytes!("../data/t2s.profile")[..]).unwrap());
 static S2T: Lazy<HashMap<String, String>> =
     Lazy::new(|| deserialize_from(&include_bytes!("../data/s2t.profile")[..]).unwrap());
+
+static T2S_FST: Lazy<Fst<Vec<u8>>> = Lazy::new(|| {
+    let mut keys: Vec<_> = T2S.keys().collect();
+    keys.sort_unstable();
+    Fst::from_iter_set(keys).unwrap()
+});
+
+static S2T_FST: Lazy<Fst<Vec<u8>>> = Lazy::new(|| {
+    let mut keys: Vec<_> = S2T.keys().collect();
+    keys.sort_unstable();
+    Fst::from_iter_set(keys).unwrap()
+});
 
 fn is_script(
     raw: &str,
@@ -65,31 +78,25 @@ pub fn is_simplified(raw: &str) -> bool {
     is_script(raw, &S2T, &T2S)
 }
 
-fn convert_script(raw: &str, mapping: &HashMap<String, String>, max: usize) -> String {
+fn convert_script(raw: &str, mapping: &HashMap<String, String>, fst: &Fst<Vec<u8>>) -> String {
     let mut converted_characters: String = String::new();
-    let default_take = raw.chars().take(max).count();
     let mut skip_bytes = 0;
-    let mut take = default_take;
 
     while skip_bytes < raw.len() {
-        let (index, last_char) = raw[skip_bytes..].char_indices().take(take).last().unwrap();
-        let end_index = index + last_char.len_utf8();
-        let substr = &raw[skip_bytes..][..end_index];
+        let tailstr = &raw[skip_bytes..];
 
-        match mapping.get(substr) {
-            Some(mapped_char) => {
-                converted_characters.push_str(mapped_char);
-                skip_bytes += substr.len();
-                take = default_take;
+        match find_longest_prefix(fst, tailstr.as_bytes()) {
+            Some((_, length)) => {
+                let tailstr = &tailstr[..length];
+                // TODO remove unwrap
+                let mapped = mapping.get(tailstr).unwrap();
+                converted_characters.push_str(mapped);
+                skip_bytes += tailstr.len();
             }
             None => {
-                if take > 1 {
-                    take -= 1;
-                } else {
-                    converted_characters.push_str(substr);
-                    skip_bytes += substr.len();
-                    take = default_take;
-                }
+                let first = tailstr.chars().next().unwrap();
+                converted_characters.push(first);
+                skip_bytes += first.len_utf8();
             }
         }
     }
@@ -98,11 +105,41 @@ fn convert_script(raw: &str, mapping: &HashMap<String, String>, max: usize) -> S
 }
 
 pub fn traditional_to_simplified(raw: &str) -> String {
-    convert_script(raw, &T2S, 20)
+    convert_script(raw, &T2S, &T2S_FST)
 }
 
 pub fn simplified_to_traditional(raw: &str) -> String {
-    convert_script(raw, &S2T, 20)
+    convert_script(raw, &S2T, &S2T_FST)
+}
+
+/// Thanks to @llogiq for this function
+/// https://github.com/BurntSushi/fst/pull/104/files
+///
+/// find the longest key that is prefix of the given value.
+///
+/// If the key exists, then `Some((value, key_len))` is returned, where
+/// `value` is the value associated with the key, and `key_len` is the
+/// length of the found key. Otherwise `None` is returned.
+///
+/// This can be used to e.g. build tokenizing functions.
+#[inline]
+fn find_longest_prefix(fst: &Fst<Vec<u8>>, value: &[u8]) -> Option<(u64, usize)> {
+    let mut node = fst.root();
+    let mut out = Output::zero();
+    let mut last_match = None;
+    for (i, &b) in value.iter().enumerate() {
+        if let Some(trans_index) = node.find_input(b) {
+            let t = node.transition(trans_index);
+            node = fst.node(t.addr);
+            out = out.cat(t.out);
+            if node.is_final() {
+                last_match = Some((out.cat(node.final_output()).value(), i + 1));
+            }
+        } else {
+            return last_match;
+        }
+    }
+    last_match
 }
 
 #[cfg(test)]
@@ -130,12 +167,20 @@ mod tests {
         let simplified = "欧洲";
         let traditional = "歐洲";
         assert_eq!(simplified, traditional_to_simplified(traditional));
+
+        let traditional = "人人生而自由﹐在尊嚴和權利上一律平等。他們賦有理性和良心﹐並應以兄弟關係的精神互相對待。";
+        let simplified = "人人生而自由﹐在尊严和权利上一律平等。他们赋有理性和良心﹐并应以兄弟关系的精神互相对待。";
+        assert_eq!(simplified, traditional_to_simplified(traditional));
     }
 
     #[test]
     fn test_simplified_to_traditional() {
         let simplified = "欧洲";
         let traditional = "歐洲";
+        assert_eq!(traditional, simplified_to_traditional(simplified));
+
+        let traditional = "人人生而自由﹐在尊嚴咊權利上一律平等。他們賦有理性咊良心﹐並應㕥兄弟關係的精神互相對待。";
+        let simplified = "人人生而自由﹐在尊严和权利上一律平等。他们赋有理性和良心﹐并应以兄弟关系的精神互相对待。";
         assert_eq!(traditional, simplified_to_traditional(simplified));
     }
 }