Add benchmarks and optimize character-converter.

## Commits

### Add Benchmarks
```
  test benches::simplified_to_traditional ... bench:     129,776 ns/iter (+/- 4,669)
  test benches::traditional_to_simplified ... bench:     121,232 ns/iter (+/- 2,866)
```

### Avoid string allocations
```
  test benches::simplified_to_traditional ... bench:     112,347 ns/iter (+/- 5,848)
  test benches::traditional_to_simplified ... bench:     103,843 ns/iter (+/- 6,099)
```

### Avoid iterating from the start of the string
```
  test benches::simplified_to_traditional ... bench:       94,790 ns/iter (+/- 625)
  test benches::traditional_to_simplified ... bench:       85,548 ns/iter (+/- 2,355)
```

### Create slices instead of collecting chars into an allocated string
```
  test benches::simplified_to_traditional ... bench:      19,382 ns/iter (+/- 224)
  test benches::traditional_to_simplified ... bench:      16,697 ns/iter (+/- 358)
```

### Find by prefix using an FST
```
  test benches::bench_simplified_to_traditional ... bench:       2,590 ns/iter (+/- 135)
  test benches::bench_traditional_to_simplified ... bench:       2,568 ns/iter (+/- 21)
```

### Add more benchmarks
```
  test benches::bench_simplified_is_simplified   ... bench:       2,086 ns/iter (+/- 47)
  test benches::bench_simplified_is_traditional  ... bench:         535 ns/iter (+/- 27)
  test benches::bench_simplified_to_simplified   ... bench:       2,454 ns/iter (+/- 18)
  test benches::bench_simplified_to_traditional  ... bench:       2,578 ns/iter (+/- 20)
  test benches::bench_traditional_is_simplified  ... bench:         546 ns/iter (+/- 25)
  test benches::bench_traditional_is_traditional ... bench:       1,961 ns/iter (+/- 20)
  test benches::bench_traditional_to_simplified  ... bench:       2,547 ns/iter (+/- 26)
  test benches::bench_traditional_to_traditional ... bench:       2,432 ns/iter (+/- 29)
```

### Use cow instead of always allocating string
This will avoid to create a string when no characters changes.
```diff
  test benches::bench_simplified_is_simplified   ... bench:       2,141 ns/iter (+/- 77)
  test benches::bench_simplified_is_traditional  ... bench:         550 ns/iter (+/- 37)
+ test benches::bench_simplified_to_simplified   ... bench:       2,262 ns/iter (+/- 12)
  test benches::bench_simplified_to_traditional  ... bench:       2,580 ns/iter (+/- 31)
  test benches::bench_traditional_is_simplified  ... bench:         572 ns/iter (+/- 10)
  test benches::bench_traditional_is_traditional ... bench:       2,009 ns/iter (+/- 19)
  test benches::bench_traditional_to_simplified  ... bench:       2,562 ns/iter (+/- 18)
+ test benches::bench_traditional_to_traditional ... bench:       2,210 ns/iter (+/- 38)
```

### Encode in a buffer instead of creating a string in is_script
```diff
+ test benches::bench_simplified_is_simplified   ... bench:         845 ns/iter (+/- 32)
+ test benches::bench_simplified_is_traditional  ... bench:         188 ns/iter (+/- 1)
  test benches::bench_simplified_to_simplified   ... bench:       2,260 ns/iter (+/- 18)
  test benches::bench_simplified_to_traditional  ... bench:       2,571 ns/iter (+/- 34)
+ test benches::bench_traditional_is_simplified  ... bench:         170 ns/iter (+/- 2)
+ test benches::bench_traditional_is_traditional ... bench:         787 ns/iter (+/- 6)
  test benches::bench_traditional_to_simplified  ... bench:       2,561 ns/iter (+/- 44)
  test benches::bench_traditional_to_traditional ... bench:       2,211 ns/iter (+/- 21)
```

### Use String::new() instead of to_string()
```diff
  test benches::bench_simplified_is_simplified   ... bench:         884 ns/iter (+/- 22)
  test benches::bench_simplified_is_traditional  ... bench:         194 ns/iter (+/- 4)
+ test benches::bench_simplified_to_simplified   ... bench:       2,230 ns/iter (+/- 27)
+ test benches::bench_simplified_to_traditional  ... bench:       2,532 ns/iter (+/- 19)
  test benches::bench_traditional_is_simplified  ... bench:         206 ns/iter (+/- 1)
  test benches::bench_traditional_is_traditional ... bench:         833 ns/iter (+/- 40)
+ test benches::bench_traditional_to_simplified  ... bench:       2,496 ns/iter (+/- 17)
+ test benches::bench_traditional_to_traditional ... bench:       2,212 ns/iter (+/- 32)
```

### Use with_capacity instead of new when allocating the string
```diff
  test benches::bench_simplified_is_simplified   ... bench:         889 ns/iter (+/- 23)
  test benches::bench_simplified_is_traditional  ... bench:         190 ns/iter (+/- 1)
  test benches::bench_simplified_to_simplified   ... bench:       2,235 ns/iter (+/- 10)
+ test benches::bench_simplified_to_traditional  ... bench:       2,420 ns/iter (+/- 72)
  test benches::bench_traditional_is_simplified  ... bench:         197 ns/iter (+/- 3)
  test benches::bench_traditional_is_traditional ... bench:         871 ns/iter (+/- 32)
+ test benches::bench_traditional_to_simplified  ... bench:       2,399 ns/iter (+/- 22)
  test benches::bench_traditional_to_traditional ... bench:       2,177 ns/iter (+/- 17)
```

### Use contains_key in is_script
`contains_key` version expresses better the behavior of the code despite the small performance decrease.
```diff
- test benches::bench_simplified_is_simplified   ... bench:       1,001 ns/iter (+/- 31)
- test benches::bench_simplified_is_traditional  ... bench:         217 ns/iter (+/- 3)
  test benches::bench_simplified_to_simplified   ... bench:       2,237 ns/iter (+/- 18)
  test benches::bench_simplified_to_traditional  ... bench:       2,425 ns/iter (+/- 22)
- test benches::bench_traditional_is_simplified  ... bench:         230 ns/iter (+/- 4)
- test benches::bench_traditional_is_traditional ... bench:         979 ns/iter (+/- 72)
  test benches::bench_traditional_to_simplified  ... bench:       2,406 ns/iter (+/- 56)
  test benches::bench_traditional_to_traditional ... bench:       2,189 ns/iter (+/- 20)
```

poke @kerollmops