@ManyTheFish 2022-05-04
Hey @sotch-pr35mac, I made the changes about creating the FST at compile time. Could you retry your initialization tests, please? 😊