@sotch-pr35mac 2022-05-04
Thanks for the quick turnaround, I will give it another look after work today.